requirements.txt
python-dateutil~=2.8.2
pandas~=1.4.3
numpy~=1.23.2
matplotlib~=3.5.3
ibond~=0.0.0
marshmallow~=3.17.0
setuptools~=60.2.0
Werkzeug <= 2.1.2
flask == 2.1.2
flask-restx >= 0.5.1
pytest~=7.1.2